chore: strict compiler with 13 flags — zero warnings
Add -Wall -Wextra -Werror -Wshadow -Wconversion -Wwrite-strings
-Wstrict-prototypes -Wmissing-declarations -Wmissing-prototypes
-Wnull-dereference -Wformat=2 -Wswitch-default -Wcast-qual
-Wfloat-equal to CMakeLists.txt (component only, not IDF).

Fix all warnings surfaced by strict mode (15 bugs):
  -Wcast-qual:  const-correct tbft_ar_slice_const accessor
  -Wshadow:     renamed body_size→nv_body_size, sig→nv_sig,
                proofs_size→nv_proofs_size
  -Wconversion: explicit (size_t)/(uint64_t) casts in sizeof
                arithmetic and signed/unsigned parameter passing

Note: -Wundef and -Wpedantic omitted because ESP-IDF headers
(all 40+ of them) use #if UNDEFINED_SOC_FEATURE and GCC
extensions (#include_next, ##__VA_ARGS__).

+++ src/tbft_itimer.c
@@ -53,7 +53,7 @@ void tbft_itimer_start(tbft_itimer_t *t, int64_t timeout_us)
                                  * with callback running=t->false */
     t->period_us = timeout_us;
 
-    esp_err_t err = esp_timer_start_once(t->handle, timeout_us);
+    esp_err_t err = esp_timer_start_once(t->handle, (uint64_t)timeout_us);
     if (err != ESP_OK) {
         t->running = false;
         ESP_LOGE(TAG, "esp_timer_start_once failed: %s", esp_err_to_name(err));
